# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Google Analytics Data API Wrapper - Dynamic Applications",
        "description": "API untuk mengambil dan menyajikan data dari Google Analytics 4 (GA4) Data API. Mendukung multiple aplikasi yang dikonfigurasi secara dinamis melalui database.",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://api-laravel-production-39d6.up.railway.app/api",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/aplikasi": {
            "get": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Mendapatkan daftar semua aplikasi",
                "description": "Mengambil daftar semua aplikasi yang terdaftar beserta konfigurasi analytics-nya",
                "operationId": "getAplikasiList",
                "parameters": [
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filter hanya aplikasi yang aktif",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar aplikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar aplikasi berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_aplikasi": {
                                                        "type": "string",
                                                        "example": "Aplikasi Lapakami"
                                                    },
                                                    "key_aplikasi": {
                                                        "type": "string",
                                                        "example": "lapakami"
                                                    },
                                                    "property_id": {
                                                        "type": "string",
                                                        "example": "123456789"
                                                    },
                                                    "page_path_filter": {
                                                        "type": "string",
                                                        "example": "/"
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "Aplikasi untuk laporan pengaduan"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "konfigurasi_tambahan": {
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Membuat aplikasi baru",
                "description": "Menambahkan aplikasi baru ke dalam sistem",
                "operationId": "createAplikasi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_aplikasi",
                                    "key_aplikasi",
                                    "property_id"
                                ],
                                "properties": {
                                    "nama_aplikasi": {
                                        "type": "string",
                                        "example": "Aplikasi Dashboard Baru"
                                    },
                                    "key_aplikasi": {
                                        "type": "string",
                                        "example": "dashboard_baru"
                                    },
                                    "property_id": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "page_path_filter": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Aplikasi dashboard untuk monitoring"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "konfigurasi_tambahan": {
                                        "type": "object",
                                        "example": {
                                            "key": "value"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aplikasi berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aplikasi berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aplikasi/{id}": {
            "get": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Mendapatkan detail aplikasi",
                "description": "Mengambil detail aplikasi berdasarkan ID",
                "operationId": "getAplikasiDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID aplikasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail aplikasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aplikasi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aplikasi tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Update aplikasi",
                "description": "Mengupdate data aplikasi yang sudah ada",
                "operationId": "updateAplikasi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID aplikasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_aplikasi": {
                                        "type": "string"
                                    },
                                    "key_aplikasi": {
                                        "type": "string"
                                    },
                                    "property_id": {
                                        "type": "string"
                                    },
                                    "page_path_filter": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "konfigurasi_tambahan": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aplikasi berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Hapus aplikasi",
                "description": "Menghapus aplikasi dari sistem",
                "operationId": "deleteAplikasi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID aplikasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aplikasi berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aplikasi berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aplikasi/key/{key}": {
            "get": {
                "tags": [
                    "Aplikasi Management"
                ],
                "summary": "Mendapatkan aplikasi berdasarkan key",
                "description": "Mengambil data aplikasi berdasarkan key_aplikasi",
                "operationId": "getAplikasiByKey",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key aplikasi (contoh: lapakami, dashboard)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aplikasi ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aplikasi tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}